1-API: Move exceptions into `Shared` folder
