+++ Documents/project-structure.md
@@ -13,17 +13,21 @@ src/Features/Feature
              '--Utils
  
 src/Infrastructure
-    |--ApiResponse: response DTOs shared across various endpoints
     |--Database: DB configuration, migrations, seeds
     |--Endpoints
-    |--Exceptions
     |--Extensions: extension methods to configure services & middleware pipeline
     |--HealthCheck
     '--Logs
+
+src/Shared
+    |--Dtos
+    |--Exceptions
+    '--Guards
 ```
 In the future, it may make sense to separate API logic from business logic. In this case, I
 should keep code for routing, endpoints, and application configuration in an Api project and
 move business logic stuff into a Core or Application project.
+
 
 ## API Testing projects
 - Okane.Api.Tests.Unit/Integration/E2E
@@ -43,6 +47,7 @@ src/shared
     |--components
     '--mirror the feature folder structure above...
 ```
+
 
 ### Frontend tests
 Tests will be placed by the associated source file.
