+++ Okane.Api/Shared/Exceptions/ApiException.cs
